Add proxy module and remove Vert.x from CLI
Create a separate proxy module (incus-spawn-proxy) that contains
MitmProxy and a standalone ProxyMain entry point. The CLI module no
longer depends on quarkus-vertx or Netty — its ProxyStartCommand now
launches the proxy binary as a subprocess.

Module dependency separation:
- common: jackson + quarkus-core (697 tests)
- cli: common + aesh + tamboui, NO Vert.x (146 tests)
- proxy: common + quarkus-vertx + brotli4j

ProxyService detects isx-proxy alongside isx and uses it for both
the systemd unit and macOS launchd plist, falling back to
`isx proxy start` when the separate binary is not installed.

Update install.sh and CI workflow to build and install both binaries.
