+++ cli/src/main/resources/application.properties
@@ -1,9 +1,6 @@
 quarkus.banner.enabled=false
 quarkus.log.level=WARN
 quarkus.log.console.stderr=true
-quarkus.vertx.event-loops-pool-size=2
-quarkus.vertx.caching=false
-quarkus.vertx.classpath-resolving=false
 
 # Native image config
 quarkus.ssl.native=true
