+++ pom.xml
@@ -32,6 +32,7 @@
     <modules>
         <module>common</module>
         <module>cli</module>
+        <module>proxy</module>
     </modules>
 
     <properties>
