+++ proxy/src/main/resources-filtered/build.properties
@@ -1,0 +1,1 @@
+app.version=${project.version}
