+++ proxy/src/main/resources/application.properties
@@ -1,0 +1,13 @@
+quarkus.banner.enabled=false
+quarkus.log.level=WARN
+quarkus.log.console.stderr=true
+quarkus.vertx.event-loops-pool-size=2
+quarkus.vertx.caching=false
+quarkus.vertx.classpath-resolving=false
+
+# Native image config
+quarkus.ssl.native=true
+quarkus.native.auto-service-loader-registration=false
+quarkus.native.additional-build-args=--initialize-at-run-time=dev.incusspawn.Environment,\
+  -H:-ParseRuntimeOptions,-H:-UsePerfData,-H:-UseContainerSupport,-H:-FlightRecorder,\
+  -R:MaxRAM=256m,-R:ActiveProcessorCount=2
